+++ libs/ngrx/src/lib/adapt-ngrx.service.ts
@@ -1,9 +1,9 @@
 import { Injectable } from '@angular/core';
 import { Store } from '@ngrx/store';
-import { AdaptCommon } from '@state-adapt/rxjs';
+import { Adapt } from '@state-adapt/rxjs';
 
 @Injectable({ providedIn: 'root' })
-export class AdaptNgrx extends AdaptCommon<Store> {
+export class AdaptNgrx extends Adapt<Store> {
   constructor(private store: Store) {
     super(store);
   }