+++ libs/ngxs/src/lib/adapt-ngxs.function.ts
@@ -1,8 +1,8 @@
 import { inject } from '@angular/core';
-import { AdaptCommon } from '@state-adapt/rxjs';
+import { Adapt } from '@state-adapt/rxjs';
 import { AdaptNgxs } from './adapt-ngxs.service';
 
-export const adaptNgxs: AdaptCommon['init'] = <T extends any[]>(...args: T) => {
+export const adaptNgxs: Adapt['init'] = <T extends any[]>(...args: T) => {
   const adaptDep = inject(AdaptNgxs); // Needs it from NgRx store
   return (adaptDep.init as any)(...args);
 };