+++ libs/ngxs/src/lib/adapt-ngxs.service.ts
@@ -1,9 +1,9 @@
 import { Injectable } from '@angular/core';
 import { Store } from '@ngxs/store';
-import { AdaptCommon } from '@state-adapt/rxjs';
+import { Adapt } from '@state-adapt/rxjs';
 
 @Injectable({ providedIn: 'root' })
-export class AdaptNgxs extends AdaptCommon<Store> {
+export class AdaptNgxs extends Adapt<Store> {
   constructor(private store: Store) {
     super(store);
   }