+++ libs/rxjs/src/lib/global-store/create-state-adapt.funciton.ts
@@ -1,4 +1,4 @@
-import { AdaptCommon } from './adapt';
+import { Adapt } from './adapt';
 import { AdaptStore, ObservableStore } from './adapt.store';
 
 export function createStateAdapt<
@@ -7,5 +7,5 @@ export function createStateAdapt<
   Store extends ObservableStore<State, AnyAction>,
 >(store: Store) {
   const adaptStore = new AdaptStore<any, AnyAction, any, Store>(store);
-  return new AdaptCommon(adaptStore);
+  return new Adapt(adaptStore);
 }